Keyboard name should be "Italian".
Keyboard name should be "Italian". Currently when having English keyboard by default and then adding Italian language, in the space bar appears the text "Italian keyboard" instead of "Italian". This is particularly annoying when opening the keyboard when browsing the web, because the space bar is smaller and the font size makes the text difficult to read.